Update README.md
The weather probe RS41 can only work on one 1.5V battery and that via the upconverter that is integrated on the pcb. It can work up to 1.1V battery voltage. Operation with only one battery lasts as long as with two batteries. The only thing is that it does not respond to the on/off button, but as soon as the battery is connected, the probe starts working immediately and cannot be turned off.